+++ source/detail/number_format/number_formatter.cpp
@@ -83,7 +83,7 @@ void unhandled_case(bool error)
 namespace xlnt {
 namespace detail {
 
-bool format_condition::satisfied_by(long double number) const
+bool format_condition::satisfied_by(double number) const
 {
     switch (type)
     {
@@ -1075,7 +1075,7 @@ number_formatter::number_formatter(const std::string &format_string, xlnt::calen
     format_ = parser_.result();
 }
 
-std::string number_formatter::format_number(long double number)
+std::string number_formatter::format_number(double number)
 {
     if (format_[0].has_condition)
     {
@@ -1154,7 +1154,7 @@ std::string number_formatter::format_text(const std::string &text)
     return format_text(format_[3], text);
 }
 
-std::string number_formatter::fill_placeholders(const format_placeholders &p, long double number)
+std::string number_formatter::fill_placeholders(const format_placeholders &p, double number)
 {
     std::string result;
 
@@ -1230,7 +1230,7 @@ std::string number_formatter::fill_placeholders(const format_placeholders &p, lo
     else if (p.type == format_placeholders::placeholders_type::fractional_part)
     {
         auto fractional_part = number - integer_part;
-        result = std::fabs(fractional_part) < std::numeric_limits<long double>::min()
+        result = std::fabs(fractional_part) < std::numeric_limits<double>::min()
             ? std::string(".")
             : std::to_string(fractional_part).substr(1);
 
@@ -1259,7 +1259,7 @@ std::string number_formatter::fill_placeholders(const format_placeholders &p, lo
 }
 
 std::string number_formatter::fill_scientific_placeholders(const format_placeholders &integer_part,
-    const format_placeholders &fractional_part, const format_placeholders &exponent_part, long double number)
+    const format_placeholders &fractional_part, const format_placeholders &exponent_part, double number)
 {
     std::size_t logarithm = 0;
 
@@ -1312,7 +1312,7 @@ std::string number_formatter::fill_scientific_placeholders(const format_placehol
 }
 
 std::string number_formatter::fill_fraction_placeholders(const format_placeholders & /*numerator*/,
-    const format_placeholders &denominator, long double number, bool /*improper*/)
+    const format_placeholders &denominator, double number, bool /*improper*/)
 {
     auto fractional_part = number - static_cast<int>(number);
     auto original_fractional_part = fractional_part;
@@ -1337,7 +1337,7 @@ std::string number_formatter::fill_fraction_placeholders(const format_placeholde
     {
         auto numerator_full = original_fractional_part * i;
         auto numerator_rounded = static_cast<int>(std::round(numerator_full));
-        auto difference = std::fabs(original_fractional_part - (numerator_rounded / static_cast<long double>(i)));
+        auto difference = std::fabs(original_fractional_part - (numerator_rounded / static_cast<double>(i)));
 
         if (difference < best_difference)
         {
@@ -1350,7 +1350,7 @@ std::string number_formatter::fill_fraction_placeholders(const format_placeholde
     return std::to_string(numerator_rounded) + "/" + std::to_string(best_denominator);
 }
 
-std::string number_formatter::format_number(const format_code &format, long double number)
+std::string number_formatter::format_number(const format_code &format, double number)
 {
     static const std::vector<std::string> *month_names = new std::vector<std::string>{"January", "February", "March",
         "April", "May", "June", "July", "August", "September", "October", "November", "December"};
